I updated DataJuggler.Net to version 5.6.0, as now DataTier.Net can work with Dot Net Core / Blazor.
In this release I fixed an issue where the Delete private variable and Property were being created for Views.

I also updated the depency for DataJuggler.Core.UltimateHelper to version 1.3.5, which includes a new EnvironmentVariableHelper.

I am preparing a DataTier.Net release to 1.3.0 as soon as some intiial testing passes.